Fixing a few bugs

+++ src/ssh/azext_ssh/file_utils.py
@@ -27,13 +27,15 @@ def mkdir_p(path):
 
 
 def delete_file(file_path, message, warning=False):
-    try:
-        os.remove(file_path)
-    except Exception as e:
-        if warning:
-            logger.warning(message)
-        else:
-            raise azclierror.FileOperationError(message + "Error: " + str(e)) from e
+    if os.path.isfile(file_path):
+        try:
+            os.remove(file_path)
+        except Exception as e:
+            if warning:
+                logger.warning(message)
+            else:
+                raise azclierror.FileOperationError(message + "Error: " + str(e)) from e
+
 
 
 def delete_folder(dir_path, message, warning=False):
